requirements.txt
aiofiles==0.4.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.7
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
dataclasses==0.6
decorator==4.4.0
deprecation==2.0.7
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
h11==0.9.0
h2==3.1.1
hpack==3.0.0
Hypercorn==0.5.4
hyperframe==5.2.0
idna==2.8
importlib-metadata==0.19
ipython==7.7.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jwt==0.6.1
MarkupSafe==1.1.1
mock==3.0.5
more-itertools==7.2.0
multidict==4.5.2
numpy==1.17.0
packaging==19.1
pandas==0.25.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pkg-resources==0.0.0
pluggy==0.12.0
polling==0.3.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
pyEX==0.1.15
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-engineio==3.9.3
python-socketio==4.3.1
pytoml==0.1.21
pytz==2019.2
requests==2.22.0
six==1.12.0
socketIO-client==0.7.2
socketIO-client-nexus==0.7.6
sortedcontainers==2.1.0
SQLAlchemy==1.3.7
sseclient==0.0.24
traitlets==4.3.2
typing-extensions==3.7.4
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.5
wsproto==0.15.0
zipp==0.5.2
